Merge #328, #320 and #269 (#329)
This is a PR that merges #328, #320 and #269

For some reason our CI doesn't trigger `buildkite` when coming from external repo. And I don't have superpower to merge them.

This PR consolidates 3 mentioned PRs.

#328 
--
PR by @sept-en

Currently, we are receiving just _"The account is not registered"_ error when there's no registered account and these functions of FT are called:
* `FungibleToken::internal_deposit()`
* `FungibleToken::internal_withdraw()`
* `StorageManagement::storage_deposit()`
* `StorageManagement::storage::withdraw()`

This is not very convenient and could be sometimes not that easy to find out (especially having complex contracts with cross-contracts calls) which account doesn't exist indeed. We recently ran into this problem when writing some sim tests and spent a lot amount of time trying to track the issue. Making these messages more descriptive by specifying the account id solves the issue.

#320
--
PR by @joshuajbouw

For the evm tests we require the ability to produce blocks from the account, which wraps the runtime. It would be fantastic if it is possible to get a borrowed `Ref` or `MutRef` so that we can do such.

What this does is exactly that, with the bonus of fixing a potential issue with `Runtime::current_block` which had returned a mutable `Block` and not a referenced `Block`. I'm sure it is ideal for there not to be mutable access to it as it is generated.

#269
--
PR by @luciotato

clarifying note, where does the xxxxxxxContract struct comes from in sim testing